requirements.txt
pygame
stockfish
chess
pygame_widgets
pyperclip
pygame-textinput
Pillow